+++ client/src/components/Profile/Profile.js
@@ -2,7 +2,7 @@ import React from 'react'
 import TopHeader from '../../containers/TopHeaderContainer'
 import Footer from '../Footer/Footer'
 import styled from 'styled-components'
-import { Link } from 'react-router-dom'
+import CourseCard from './CourseCard'
 import defaultAvatar from '../../assets/avatarIcon.svg'
 
 const Wrap = styled.div`
@@ -101,46 +101,6 @@ const CourseListWrap = styled.div`
   }
 `
 
-const CourseCard = styled(Link)`
-margin: 1em;
-width: 100%;
-flex-grow: 0;
-box-shadow: 0px 1px 6px rgba(0, 0, 0, 0.25);
-background-color: white;
--webkit-transition: all 450ms ease;
-transition: all 450ms ease;
-text-align: left;
-text-decoration: none;
-color: rgb(76, 87, 101);
-
-img {
-  width: 100%;
-  display: block;
-}
-
-p {
-  border-top: 1px solid rgb(226, 226, 226);
-  margin: 0;
-  padding: 15px;
-  font-size: 17px;
-}
-
-span {
-  font-size: 14px;
-  padding:5px 0 5px 15px;
-  display: inline-block;
-  font-weight: 200;
-}
-
-@media (min-width: 600px) {
-  width: calc(50% - 4em)
-}
-
-@media (min-width: 1024px) {
-  width: calc(33.33333% - 2em)
-}
-`
-
 export default ({ paidCourses, latestExpireDate, total, status, phoneNum }) => {
   switch (status) {
     case 'LOADING': {
@@ -171,11 +131,13 @@ export default ({ paidCourses, latestExpireDate, total, status, phoneNum }) => {
               ? (
                 <CourseListWrap>
                   {paidCourses.map((item, i) => (
-                    <CourseCard to={`${item.link}`} key={item.key}>
-                      <span>{item.publishedAt && item.publishedAt.substr(0, 10)}</span>
-                      <img src={`${item.cover}`} alt='cover' className='cover' />
-                      <p>{item.title}</p>
-                    </CourseCard>
+                    <CourseCard
+                      link={item.link}
+                      key={item.key}
+                      publishedAt={item.publishedAt}
+                      cover={item.cover}
+                      title={item.title}
+                    />
                 ))}
                 </CourseListWrap>
             )
